# src/main/resources/openapi3_1.yaml
openapi: 3.1.0
info:
  title: New Spa
  description: |-
    Specification for New Spa Application based on OpenAPI 3.1 specification.  
  contact:
    email: josue270193@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /user/login:
    post:
      tags:
        - user
      summary: Login user into the system
      description: ''
      operationId: loginUser
      requestBody:
        description: Obtain a token from the user logged
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Invalid username/password supplied
  /user/register:
    post:
      tags:
        - user
      summary: Create a new user from request
      description: ''
      operationId: registerUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation exception
components:
  schemas:
...
    Status:
      type: object
      properties:
        id:
          type: integer
          format: int64
          examples: [ 10 ]
        name:
          type: string
          examples: [ ACTIVE, INACTIVE ]
        entity:
          type: string
          examples: [ USER, TASK ]
      xml:
        name: status
    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          examples: [ josue270193 ]
        password:
          type: string
          format: password
          examples: [ '12345' ]
      xml:
        name: userLogin
    UserRegister:
      type: object
      required:
        - username
        - password
        - name
        - email
      properties:
        username:
          type: string
          examples: [ test ]
        password:
          type: string
          format: password
          examples: [ test ]
        name:
          type: string
          examples: [ test_name ]
        email:
          type: string
          format: email
          examples: [ test@gmail.com ]
      xml:
        name: userRegister
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          examples: [ 10 ]
        username:
          type: string
          examples: [ test ]
        password:
          type: string
          examples: [ test ]
        name:
          type: string
          examples: [ test_name ]
        email:
          type: string
          examples: [ test@gmail.com ]
        status:
          type: object
          description: the status of the user
          $ref: '#/components/schemas/Status'
      xml:
        name: user
...
  securitySchemes:
    new_spa_auth_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT